PTSE-110 check-in/check-out soon working + jest tests passing + file renamed
