+++ src/utils/reservationpropertyunifier.ts
@@ -88,7 +88,21 @@ export const getPropertiesForPropertyTable = (
     const currTime = new Date()
     let closestReservation: IReservation | undefined = undefined
     prop.reservations.forEach((r) => {
+      // if reservation is ongoing, status is either CHECK-OUT SOON (end_datetime is within 1 day reach)
+      // or OCCUPIED, if the previous condition is not met
       if (r.begin_datetime < new Date() && r.end_datetime > new Date()) {
+        if (isDateWithinNextDay(r.end_datetime)) {
+          propertyList.push({
+            id: prop._id,
+            title: prop.title,
+            address: prop.address,
+            status: PropertyStatus.CHECK_OUT_SOON,
+            arrival: r.begin_datetime,
+            departure: r.end_datetime,
+            price: prop.price,
+          })
+          return
+        }
         propertyList.push({
           id: prop._id,
           title: prop.title,
@@ -109,16 +123,31 @@ export const getPropertiesForPropertyTable = (
       )
         closestReservation = r
     })
-    // if the property is not occupied but has some reservations,
-    // add it as "Free", but with the closest reservation's dates as arrival and departure time
+    // if the property is not yet in propertyList after iterating all reservations, then it either:
+    //    1. has no reservations in the future, only the past -> arrival/dep undefined
+    //    2. has no current reservations, but has reservations in the future -> arrival/dep based on closestReservation
+    // its status will be either CHECK-IN SOON, if closestReservation is within 1 day reach
+    // or FREE, if closestReservation is not within 1 day reach
     if (!propertyList.find((p) => p.id === prop._id)) {
+      if (closestReservation && isDateWithinNextDay((closestReservation as IReservation).begin_datetime)) {
+        propertyList.push({
+          id: prop._id,
+          title: prop.title,
+          address: prop.address,
+          status: PropertyStatus.CHECK_IN_SOON,
+          arrival: (closestReservation as IReservation).begin_datetime,
+          departure: (closestReservation as IReservation).end_datetime,
+          price: prop.price,
+        })
+        return
+      }
       propertyList.push({
         id: prop._id,
         title: prop.title,
         address: prop.address,
         status: PropertyStatus.FREE,
-        arrival: closestReservation !== undefined ? closestReservation.begin_datetime : undefined,
-        departure: closestReservation !== undefined ? closestReservation.end_datetime : undefined,
+        arrival: closestReservation !== undefined ? (closestReservation as IReservation).begin_datetime : undefined,
+        departure: closestReservation !== undefined ? (closestReservation as IReservation).end_datetime : undefined,
         price: prop.price,
       })
     }
@@ -132,10 +161,10 @@ const isEventUpcoming = (event: IEvent | IReservation): boolean => {
   return event.begin_datetime.getTime() - currTime.getTime() > 0
 }
 
-const isEventInTheNextDay = (event: IEvent): boolean => {
+const isDateWithinNextDay = (date: Date): boolean => {
   const currTime = new Date();
-  const millisecondsFromEventBeginTime = event.begin_datetime.getTime() - currTime.getTime();
+  const millisecondsFromDateTime = date.getTime() - currTime.getTime();
   // returns true if event begin_time is within 1 day worth of milliseconds (86400 * 1000)
   // from starting
-  return millisecondsFromEventBeginTime < 86400 * 1000 && millisecondsFromEventBeginTime > 0;
+  return millisecondsFromDateTime < 86400 * 1000 && millisecondsFromDateTime > 0;
 }
